preventing user from editing and deleting someone else's post

+++ backend/routes/posts.js
@@ -35,6 +35,7 @@ router.post('',
       title: req.body.title,
       content: req.body.content,
       imagePath: url + '/images/' + req.file.filename,
+      creator: req.userData.userId,
     });
 
     post.save().then((createdPost) => {
@@ -63,13 +64,20 @@ router.put('/:id',
       title: req.body.title,
       content: req.body.content,
       imagePath: imagePath,
+      creator: req.userData.userId,
     });
 
-    Post.updateOne({_id: req.params.id}, post)
+    Post.updateOne({_id: req.params.id, creator: req.userData.userId}, post)
       .then(result => {
-        res.status(200).json({
-          message: 'Updated successfully!'
-        });
+        if (result.modifiedCount > 0) {
+          res.status(200).json({
+            message: 'Updated successfully!'
+          });
+        } else {
+          res.status(401).json({
+            message: 'Not authorized!'
+          });
+        }
       });
   });
 
@@ -114,8 +122,16 @@ router.get('/:id', (req, res, next) => {
 
 router.delete('/:id', checkAuth,
   (req, res, next) => {
-    Post.deleteOne({_id: req.params.id}).then(result => {
-      res.status(200).json({message: 'Post deleted!'});
+    Post.deleteOne({_id: req.params.id, creator: req.userData.userId}).then(result => {
+      if (result.deletedCount > 0) {
+        res.status(200).json({
+          message: 'Post deleted!'
+        });
+      } else {
+        res.status(401).json({
+          message: 'Not authorized!'
+        });
+      }
     });
   });
